requirements.txt
altair

numpy
pandas
pandas-datareader
Pillow
plotly
scikit-learn >= 0.18
scipy

streamlit
yfinance